backend/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
idna==3.4
numpy==1.23.4
optional-django==0.3.0
pydub==0.25.1
python-core==0.0.1
pytz==2022.6
react==4.3.0
requests==2.28.1
Serializer==0.2.1
simplejson==3.17.6
SpeechRecognition==3.8.1
sqlparse==0.4.3
urllib3==1.26.12
